requirements.txt
streamlit
pandas
numpy
openpyxl
plotly
python-docx
xlsxwriter